upload: override datetime automatically
Remove the --datetime flag to the "upload" subcommand. This was just
another knob that we should be able to determine automatically from the
files in chacra.

If we didn't have a log file (--skip-log), search for a .changes file
instead. When we find it, use the "Date" value from that.