session: use higher-level API methods
It is not always sufficient to pass in an empty opts dict "{}" to
ClientSession(). For example, GSSAPI auth with an "HTTP" krbservice
principal overrides the default krbservice setting of "host", and we
need to read this from our profile and pass this into ClientSession().

Use the higher-level Koji client API methods to get the entire profile
module and configuration.